Rename `cli_paths` -> `selectors`
